% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alkyoni}
\alias{alkyoni}
\alias{theme_roboto}
\title{telca: A package to make plotting with ggplot2 a bit easier}
\usage{
theme_roboto(
  base_size = 12,
  base_line_size = base_size/24,
  base_rect_size = base_size/24,
  base_family = "Roboto"
)
}
\arguments{
\item{base_size}{base font size}

\item{base_line_size, base_rect_size, base_family}{check#'}
}
\description{
telca: A package to make plotting with ggplot2 a bit easier
}
\details{
The Roboto Condensed and Roboto Bold fonts are both Google fonts;
they can be found at \url{https://fonts.google.com/specimen/Roboto+Condensed}
and \url{https://fonts.google.com/specimen/Roboto}. These fonts must be
installed locally on your computer for this theme to work.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?") +
    theme_roboto()
}

}
